Level 0 Titles
Added chapter group titles to match the grouping used in the
docs.operable.io book.  Used Level 0 headers in the cog_book.adoc itself
to make this level of groupings easier to review and manage in the
future.   I did not change the sidebar TOC formatting.  That will need
review.